refactor: extract story io helpers from app
