+++ tests/conftest.py
@@ -129,6 +129,36 @@ SAMPLE_NPCS = [
 ]
 
 
+def _patch_paths_all_modules(monkeypatch, tmp_path, stories_dir, design_dir, app_module=None):
+    """Patch shared path constants across app and extracted helper modules."""
+    import app as imported_app_module
+    import event_db
+    import lore_db
+    import state_db
+    import story_io
+
+    app_target = app_module or imported_app_module
+    data_dir = tmp_path / "data"
+    patches = (
+        ("BASE_DIR", str(tmp_path)),
+        ("DATA_DIR", str(data_dir)),
+        ("STORIES_DIR", str(stories_dir)),
+        ("STORY_DESIGN_DIR", str(design_dir)),
+        ("STORIES_REGISTRY_PATH", str(data_dir / "stories.json")),
+        ("_LLM_CONFIG_PATH", str(tmp_path / "llm_config.json")),
+    )
+    for mod in (app_target, story_io, event_db, lore_db, state_db):
+        for attr, value in patches:
+            if hasattr(mod, attr):
+                monkeypatch.setattr(mod, attr, value)
+    lore_db._embedding_cache.clear()
+
+
+@pytest.fixture
+def patch_paths_all_modules():
+    return _patch_paths_all_modules
+
+
 # ---------------------------------------------------------------------------
 # Fixtures
 # ---------------------------------------------------------------------------
