+++ tests/test_debug_api.py
@@ -11,22 +11,13 @@ import state_db
 
 
 @pytest.fixture(autouse=True)
-def patch_app_paths(tmp_path, monkeypatch):
+def patch_app_paths(tmp_path, monkeypatch, patch_paths_all_modules):
     data_dir = tmp_path / "data"
     stories_dir = data_dir / "stories"
     stories_dir.mkdir(parents=True)
     design_dir = tmp_path / "story_design"
     design_dir.mkdir()
-    monkeypatch.setattr(app_module, "BASE_DIR", str(tmp_path))
-    monkeypatch.setattr(app_module, "DATA_DIR", str(data_dir))
-    monkeypatch.setattr(app_module, "STORIES_DIR", str(stories_dir))
-    monkeypatch.setattr(app_module, "STORY_DESIGN_DIR", str(design_dir))
-    monkeypatch.setattr(app_module, "STORIES_REGISTRY_PATH", str(data_dir / "stories.json"))
-    monkeypatch.setattr(event_db, "STORIES_DIR", str(stories_dir))
-    monkeypatch.setattr(lore_db, "STORIES_DIR", str(stories_dir))
-    monkeypatch.setattr(lore_db, "STORY_DESIGN_DIR", str(design_dir))
-    monkeypatch.setattr(state_db, "STORIES_DIR", str(stories_dir))
-    lore_db._embedding_cache.clear()
+    patch_paths_all_modules(monkeypatch, tmp_path, stories_dir, design_dir, app_module=app_module)
     monkeypatch.setattr(app_module, "_log_llm_usage", lambda *a, **kw: None)
     return stories_dir
