+++ tests/test_dungeon_reconciliation.py
@@ -168,17 +168,13 @@ import pytest
 
 
 @pytest.fixture(autouse=True)
-def patch_paths(tmp_path, monkeypatch):
+def patch_paths(tmp_path, monkeypatch, patch_paths_all_modules):
     stories_dir = tmp_path / "data" / "stories"
     design_dir = tmp_path / "story_design"
     stories_dir.mkdir(parents=True)
     design_dir.mkdir(parents=True)
 
-    monkeypatch.setattr(app_module, "STORIES_DIR", str(stories_dir))
-    monkeypatch.setattr(app_module, "STORY_DESIGN_DIR", str(design_dir))
-    monkeypatch.setattr(app_module, "BASE_DIR", str(tmp_path))
-    monkeypatch.setattr(state_db, "STORIES_DIR", str(stories_dir))
-    monkeypatch.setattr(event_db, "STORIES_DIR", str(stories_dir))
+    patch_paths_all_modules(monkeypatch, tmp_path, stories_dir, design_dir, app_module=app_module)
     monkeypatch.setattr(world_timer, "BASE_DIR", str(tmp_path))
     monkeypatch.setattr(
         dungeon_system,
