+++ tests/test_state_cleanup.py
@@ -30,17 +30,13 @@ CLEANUP_SCHEMA = {
 
 
 @pytest.fixture(autouse=True)
-def patch_paths(tmp_path, monkeypatch):
+def patch_paths(tmp_path, monkeypatch, patch_paths_all_modules):
     """Redirect app, state_db, event_db to tmp_path."""
     stories_dir = tmp_path / "data" / "stories"
     stories_dir.mkdir(parents=True)
     design_dir = tmp_path / "story_design"
     design_dir.mkdir()
-    monkeypatch.setattr(app_module, "STORIES_DIR", str(stories_dir))
-    monkeypatch.setattr(app_module, "STORY_DESIGN_DIR", str(design_dir))
-    monkeypatch.setattr(app_module, "BASE_DIR", str(tmp_path))
-    monkeypatch.setattr(state_db, "STORIES_DIR", str(stories_dir))
-    monkeypatch.setattr(event_db, "STORIES_DIR", str(stories_dir))
+    patch_paths_all_modules(monkeypatch, tmp_path, stories_dir, design_dir, app_module=app_module)
     (design_dir / STORY_ID).mkdir(parents=True, exist_ok=True)
     (design_dir / STORY_ID / "character_schema.json").write_text(
         json.dumps(CLEANUP_SCHEMA, ensure_ascii=False), encoding="utf-8"
